[#213] Test: Category Executor 테스트를 CategoryServiceIntegrationTest로 통합
