+++ src/test/java/soon/fridgely/domain/category/service/CategoryCacheIntegrationTest.java
@@ -13,6 +13,7 @@ import soon.fridgely.domain.member.entity.Member;
 import soon.fridgely.domain.member.repository.MemberRepository;
 import soon.fridgely.domain.refrigerator.dto.command.MemberRefrigeratorKey;
 import soon.fridgely.domain.refrigerator.entity.Refrigerator;
+import soon.fridgely.domain.refrigerator.repository.MemberRefrigeratorRepository;
 import soon.fridgely.domain.refrigerator.repository.RefrigeratorRepository;
 import soon.fridgely.global.support.RedisIntegrationTestSupport;
 
@@ -20,22 +21,14 @@ import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static soon.fridgely.global.support.fixture.MemberFixture.member;
+import static soon.fridgely.global.support.fixture.MemberRefrigeratorFixture.memberRefrigerator;
 import static soon.fridgely.global.support.fixture.RefrigeratorFixture.refrigerator;
 
 @TestPropertySource(properties = "spring.cache.type=redis")
 class CategoryCacheIntegrationTest extends RedisIntegrationTestSupport {
 
     @Autowired
-    private CategoryFinder categoryFinder;
-
-    @Autowired
-    private CategoryAppender categoryAppender;
-
-    @Autowired
-    private CategoryModifier categoryModifier;
-
-    @Autowired
-    private CategoryRemover categoryRemover;
+    private CategoryService categoryService;
 
     @Autowired
     private CategoryRepository categoryRepository;
@@ -45,6 +38,9 @@ class CategoryCacheIntegrationTest extends RedisIntegrationTestSupport {
 
     @Autowired
     private RefrigeratorRepository refrigeratorRepository;
+
+    @Autowired
+    private MemberRefrigeratorRepository memberRefrigeratorRepository;
 
     @Autowired
     private CacheManager cacheManager;
@@ -60,6 +56,9 @@ class CategoryCacheIntegrationTest extends RedisIntegrationTestSupport {
         this.refrigerator = refrigeratorRepository.save(
             refrigerator(fixtureMonkey).sample()
         );
+        memberRefrigeratorRepository.save(
+            memberRefrigerator(fixtureMonkey, refrigerator, member).sample()
+        );
 
         Cache cache = cacheManager.getCache("categories");
         if (cache != null) {
@@ -71,10 +70,10 @@ class CategoryCacheIntegrationTest extends RedisIntegrationTestSupport {
     void 카테고리_목록_조회_시_캐시가_적용된다() {
         // given
         MemberRefrigeratorKey key = new MemberRefrigeratorKey(member.getId(), refrigerator.getId());
-        categoryAppender.appendDefaultCategories(key);
+        categoryService.appendDefaultCategories(key);
 
         // when
-        CachedCategories firstResult = categoryFinder.findAll(refrigerator.getId());
+        CachedCategories firstResult = categoryService.findAll(refrigerator.getId());
 
         // then
         assertThat(firstResult.categories()).hasSize(8);
@@ -82,7 +81,7 @@ class CategoryCacheIntegrationTest extends RedisIntegrationTestSupport {
         assertThat(cache).isNotNull();
         assertThat(cache.get(refrigerator.getId())).isNotNull();
 
-        CachedCategories secondResult = categoryFinder.findAll(refrigerator.getId());
+        CachedCategories secondResult = categoryService.findAll(refrigerator.getId());
         assertThat(secondResult.categories()).hasSize(8);
         assertThat(firstResult).isEqualTo(secondResult);
     }
@@ -91,17 +90,17 @@ class CategoryCacheIntegrationTest extends RedisIntegrationTestSupport {
     void 커스텀_카테고리_추가_시_캐시가_무효화된다() {
         // given
         MemberRefrigeratorKey key = new MemberRefrigeratorKey(member.getId(), refrigerator.getId());
-        categoryAppender.appendDefaultCategories(key);
+        categoryService.appendDefaultCategories(key);
 
-        CachedCategories cachedResult = categoryFinder.findAll(refrigerator.getId());
+        CachedCategories cachedResult = categoryService.findAll(refrigerator.getId());
         assertThat(cachedResult.categories()).hasSize(8);
 
         // when
         AddCategory addCategory = new AddCategory("새 카테고리", refrigerator.getId(), member.getId());
-        categoryAppender.appendCustomCategory(addCategory);
+        categoryService.appendCustomCategory(addCategory);
 
         // then
-        CachedCategories updatedResult = categoryFinder.findAll(refrigerator.getId());
+        CachedCategories updatedResult = categoryService.findAll(refrigerator.getId());
         assertThat(updatedResult.categories()).hasSize(9)
             .extracting(CachedCategoryInfo::name)
             .contains("새 카테고리");
@@ -113,13 +112,13 @@ class CategoryCacheIntegrationTest extends RedisIntegrationTestSupport {
         MemberRefrigeratorKey key = new MemberRefrigeratorKey(member.getId(), refrigerator.getId());
 
         // when
-        categoryAppender.appendDefaultCategories(key);
+        categoryService.appendDefaultCategories(key);
 
         // then
         Cache cache = cacheManager.getCache("categories");
         assertThat(cache.get(refrigerator.getId())).isNull();
 
-        CachedCategories categories = categoryFinder.findAll(refrigerator.getId());
+        CachedCategories categories = categoryService.findAll(refrigerator.getId());
         assertThat(categories.categories()).hasSize(8)
             .extracting(CachedCategoryInfo::name)
             .containsExactlyInAnyOrder("야채", "과일", "육류", "해산물", "유제품", "음료", "간식", "기타");
@@ -129,12 +128,12 @@ class CategoryCacheIntegrationTest extends RedisIntegrationTestSupport {
     void 카테고리_수정_시_캐시가_무효화된다() {
         // given
         MemberRefrigeratorKey key = new MemberRefrigeratorKey(member.getId(), refrigerator.getId());
-        categoryAppender.appendDefaultCategories(key);
+        categoryService.appendDefaultCategories(key);
 
         AddCategory addCategory = new AddCategory("수정될 카테고리", refrigerator.getId(), member.getId());
-        categoryAppender.appendCustomCategory(addCategory);
+        categoryService.appendCustomCategory(addCategory);
 
-        CachedCategories cachedCategories = categoryFinder.findAll(refrigerator.getId());
+        CachedCategories cachedCategories = categoryService.findAll(refrigerator.getId());
         CachedCategoryInfo targetCategory = cachedCategories.categories().stream()
             .filter(c -> c.name().equals("수정될 카테고리"))
             .findFirst()
@@ -147,10 +146,10 @@ class CategoryCacheIntegrationTest extends RedisIntegrationTestSupport {
             refrigerator.getId(),
             targetCategory.id()
         );
-        categoryModifier.modify(modifyCategory);
+        categoryService.modifyCustomCategory(modifyCategory);
 
         // then
-        CachedCategories updatedCategories = categoryFinder.findAll(refrigerator.getId());
+        CachedCategories updatedCategories = categoryService.findAll(refrigerator.getId());
         assertThat(updatedCategories.categories())
             .extracting(CachedCategoryInfo::name)
             .contains("수정된 카테고리")
@@ -161,11 +160,11 @@ class CategoryCacheIntegrationTest extends RedisIntegrationTestSupport {
     void 카테고리_삭제_시_캐시가_무효화된다() {
         // given
         MemberRefrigeratorKey key = new MemberRefrigeratorKey(member.getId(), refrigerator.getId());
-        categoryAppender.appendDefaultCategories(key);
+        categoryService.appendDefaultCategories(key);
         AddCategory addCategory = new AddCategory("삭제될 카테고리", refrigerator.getId(), member.getId());
-        categoryAppender.appendCustomCategory(addCategory);
+        categoryService.appendCustomCategory(addCategory);
 
-        CachedCategories cachedCategories = categoryFinder.findAll(refrigerator.getId());
+        CachedCategories cachedCategories = categoryService.findAll(refrigerator.getId());
         CachedCategoryInfo targetCategory = cachedCategories.categories().stream()
             .filter(c -> c.name().equals("삭제될 카테고리"))
             .findFirst()
@@ -179,10 +178,10 @@ class CategoryCacheIntegrationTest extends RedisIntegrationTestSupport {
             refrigerator.getId(),
             targetCategory.id()
         );
-        categoryRemover.remove(deleteCategory);
+        categoryService.removeCustomCategory(deleteCategory);
 
         // then
-        CachedCategories updatedCategories = categoryFinder.findAll(refrigerator.getId());
+        CachedCategories updatedCategories = categoryService.findAll(refrigerator.getId());
         assertThat(updatedCategories.categories())
             .extracting(CachedCategoryInfo::name)
             .doesNotContain("삭제될 카테고리");
@@ -193,15 +192,18 @@ class CategoryCacheIntegrationTest extends RedisIntegrationTestSupport {
         // given
         Member member2 = memberRepository.save(member(fixtureMonkey).sample());
         Refrigerator refrigerator2 = refrigeratorRepository.save(refrigerator(fixtureMonkey).sample());
+        memberRefrigeratorRepository.save(
+            memberRefrigerator(fixtureMonkey, refrigerator2, member2).sample()
+        );
 
         MemberRefrigeratorKey key1 = new MemberRefrigeratorKey(member.getId(), refrigerator.getId());
         MemberRefrigeratorKey key2 = new MemberRefrigeratorKey(member2.getId(), refrigerator2.getId());
 
-        categoryAppender.appendDefaultCategories(key1);
-        categoryAppender.appendDefaultCategories(key2);
+        categoryService.appendDefaultCategories(key1);
+        categoryService.appendDefaultCategories(key2);
 
-        categoryFinder.findAll(refrigerator.getId());
-        categoryFinder.findAll(refrigerator2.getId());
+        categoryService.findAll(refrigerator.getId());
+        categoryService.findAll(refrigerator2.getId());
 
         Cache cache = cacheManager.getCache("categories");
         assertThat(cache.get(refrigerator.getId())).isNotNull();
@@ -209,14 +211,14 @@ class CategoryCacheIntegrationTest extends RedisIntegrationTestSupport {
 
         // when
         AddCategory addCategory = new AddCategory("새 카테고리", refrigerator.getId(), member.getId());
-        categoryAppender.appendCustomCategory(addCategory);
+        categoryService.appendCustomCategory(addCategory);
 
         // then
         assertThat(cache.get(refrigerator.getId())).isNull();
         assertThat(cache.get(refrigerator2.getId())).isNotNull();
 
-        CachedCategories refrigerator1Categories = categoryFinder.findAll(refrigerator.getId());
-        CachedCategories refrigerator2Categories = categoryFinder.findAll(refrigerator2.getId());
+        CachedCategories refrigerator1Categories = categoryService.findAll(refrigerator.getId());
+        CachedCategories refrigerator2Categories = categoryService.findAll(refrigerator2.getId());
 
         assertThat(refrigerator1Categories.categories()).hasSize(9);
         assertThat(refrigerator2Categories.categories()).hasSize(8);
@@ -225,12 +227,12 @@ class CategoryCacheIntegrationTest extends RedisIntegrationTestSupport {
     @Test
     void 캐시_키가_refrigeratorId로_정확하게_생성된다() {
         // given
-        categoryAppender.appendDefaultCategories(
+        categoryService.appendDefaultCategories(
             new MemberRefrigeratorKey(member.getId(), refrigerator.getId())
         );
 
         // when
-        categoryFinder.findAll(refrigerator.getId());
+        categoryService.findAll(refrigerator.getId());
 
         // then
         Cache cache = cacheManager.getCache("categories");
@@ -242,12 +244,12 @@ class CategoryCacheIntegrationTest extends RedisIntegrationTestSupport {
     @Test
     void 캐시된_데이터와_DB_데이터가_일치한다() {
         // given
-        categoryAppender.appendDefaultCategories(
+        categoryService.appendDefaultCategories(
             new MemberRefrigeratorKey(member.getId(), refrigerator.getId())
         );
 
         // when
-        CachedCategories cachedResult = categoryFinder.findAll(refrigerator.getId());
+        CachedCategories cachedResult = categoryService.findAll(refrigerator.getId());
 
         // then
         List<Category> dbResult = categoryRepository.findAllByRefrigeratorIdAndStatus(
@@ -265,18 +267,17 @@ class CategoryCacheIntegrationTest extends RedisIntegrationTestSupport {
     @Test
     void 캐시_무효화_후_재조회_시_최신_데이터를_가져온다() {
         // given
-        categoryAppender.appendDefaultCategories(new MemberRefrigeratorKey(member.getId(), refrigerator.getId()));
-        categoryFinder.findAll(refrigerator.getId());
+        categoryService.appendDefaultCategories(new MemberRefrigeratorKey(member.getId(), refrigerator.getId()));
+        categoryService.findAll(refrigerator.getId());
 
         // when
-        categoryAppender.appendCustomCategory(new AddCategory("테스트 카테고리", refrigerator.getId(), member.getId()));
+        categoryService.appendCustomCategory(new AddCategory("테스트 카테고리", refrigerator.getId(), member.getId()));
 
         // then
-        CachedCategories firstResult = categoryFinder.findAll(refrigerator.getId());
-        CachedCategories secondResult = categoryFinder.findAll(refrigerator.getId());
+        CachedCategories firstResult = categoryService.findAll(refrigerator.getId());
+        CachedCategories secondResult = categoryService.findAll(refrigerator.getId());
         assertThat(firstResult.categories()).hasSize(9);
         assertThat(secondResult.categories()).hasSize(9);
         assertThat(firstResult).isEqualTo(secondResult);
     }
-
 }
